Refactor: (Feedback) 입력값 검증 로직 이동 (Validator 클래스 제거, Splitter 로 검증 메서드 이동)
