+++ src/main/java/lotto/Main.java
@@ -3,7 +3,6 @@ package lotto;
 import lotto.domain.Calculator;
 import lotto.domain.Expression;
 import lotto.util.ExpressionSplitter;
-import lotto.util.InputValidator;
 import lotto.ui.InputView;
 import lotto.ui.ResultView;
 
@@ -15,10 +14,7 @@ public class Main {
 
         try {
             String expressionString = InputView.readExpression();
-            InputValidator.validateEmptyExpression(expressionString);
-
             List<String> expressionStrings = ExpressionSplitter.split(expressionString);
-            InputValidator.validateElementSize(expressionStrings.size());
 
             Calculator calculator = new Calculator(new Expression(expressionStrings));
             int result = calculator.calculate();
