+++ src/test/java/lotto/ExpressionTest.java
@@ -4,6 +4,7 @@ import lotto.domain.Expression;
 import lotto.exception.IllegalOperatorExpression;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
