+++ src/test/java/lotto/ExpressionValidatorTest.java
@@ -1,0 +1,17 @@
+package lotto;
+
+import lotto.exception.IllegalExpressionElementSizeException;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.NullAndEmptySource;
+
+import java.text.MessageFormat;
+
+import static org.assertj.core.api.Assertions.*;
+
+public class ExpressionValidatorTest {
+
+
+
+}
